requirements.txt
google-adk>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
telnyx>=2.1.0
googlemaps>=4.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
